search_within_pdf/requirements.txt
streamlit
pymongo
bcrypt
pytesseract
Pillow
captcha
PyMuPDF
